Extract rewards from Quest into new class QuestRewards
